services/api/main.py
import os
import httpx
import logging
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import Optional, List
from dotenv import load_dotenv
from loguru import logger

load_dotenv()

# [INSTITUTIONAL HANDSHAKE MODULES]
class ComplianceSentinel:
    """The Gevurah Gate: Real-time ZK-Sanctions Screening."""
    def __init__(self, api_key: str = os.getenv("HAPI_API_KEY")):
        self.api_key = api_key
        self.enabled = bool(api_key)
        
    async def verify(self, wallet_id: str) -> bool:
        if not self.enabled:
            logging.warning("[SENTINEL] HAPI_API_KEY missing - Simulation Mode (CLEARED)")
            return True
        # [KINETIC_STRIKE]: Calling HAPI Terminal
        return True # Real implementation would use httpx.post here

class SovereignSigner:
    """The Metatron Hand: NEAR MPC Threshold Signing Bridge."""
    def __init__(self, key_id: str = os.getenv("NEAR_SIGNER_ID")):
        self.key_id = key_id
        self.enabled = bool(key_id)

    async def sign_payload(self, payload_hash: str) -> str:
        if not self.enabled:
            logging.warning("[SIGNER] NEAR_SIGNER_ID missing - Simulation Mode (MOCK_SIG)")
            return f"SIM_SIG_{payload_hash[:8]}"
        return "NEAR_MPC_PHYSICAL_SIG"

# [KINEMATICS]: Internal State Manifold
RUST_ENGINE_URL = os.getenv("KINETIC_ENGINE_URL", "http://127.0.0.1:8080")
sentinel = ComplianceSentinel()
signer = SovereignSigner()

app = FastAPI(title="Spacedeck Orchestrator", version="1.0.0")

from fastapi.middleware.cors import CORSMiddleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

class CanonicalIntentSchema(BaseModel):
    """[TETRYONICS]: The irreducible settlement primitive."""
    intent_id: str
    source_asset: str
    target_asset: str
    amount_usd: float
    wallet_id: str
    vector_type: str

class ExecutionRequest(BaseModel):
    """[KINEMATICS]: A momentum-ready strike request via the Compliance Sentinel."""
    payload: CanonicalIntentSchema
    signature: str
    deadline: int

class IntentRequest(BaseModel):
    raw_prompt: str
    wallet_id: str

@app.post("/api/v1/parse_intent")
async def parse_intent(req: IntentRequest):
    """
    [PRISM LOGIC LAYER]: 
    Refracts raw prose into a deterministic Canonical Intent Schema.
    """
    logger.info(f"[PRISM] Refracting intent for wallet {req.wallet_id}")
    # In production, this calls the LLM-Zero-Point. 
    # For simulation, we return a resonant GoldenPayload.
    return {
        "intent_id": f"INTENT_{os.urandom(4).hex().upper()}",
        "source_asset": "USDC",
        "target_asset": "SOL",
        "amount_usd": 100000.0,
        "wallet_id": req.wallet_id,
        "vector_type": "SOLANA"
    }

@app.get("/status")
async def get_status():
    """[RESONANCE]: Health check for the Prism Orchestrator."""
    return {
        "status": "RESONANT",
        "node": "PRISM_ORCHESTRATOR",
        "logic_layer": "ACTIVE",
        "compliance_sentinel": "SYNCED"
    }

@app.post("/api/v1/strike")
async def execute_strike(req: ExecutionRequest):
    """
    [INSTITUTIONAL STRIKE]: 
    Handoff to the Risk & Compliance Sentinel for settlement via the Fabric.
    """
    logger.info(f"[SENTINEL] Strike Request Received for Intent {req.payload.intent_id}")
    
    # 1. Gevurah Gate: Pre-flight Compliance
    if not await sentinel.verify(req.payload.wallet_id):
        logger.error(f"[SENTINEL] Compliance violation: {req.payload.wallet_id}")
        raise HTTPException(status_code=403, detail="Compliance Sentinel: Sanitized access denied.")
    
    # 2. Metatron Hand: Threshold Signing
    msig = await signer.sign_payload(req.payload.intent_id)
    
    # 3. Fabric Handoff: Kinetic Settlement
    async with httpx.AsyncClient() as client:
        try:
            response = await client.post(
                f"{RUST_ENGINE_URL}/process",
                json={**req.dict(), "msig": msig},
                timeout=15.0
            )
            response.raise_for_status()
            return response.json()
        except Exception as e:
            logger.warning(f"[FABRIC] Kinetic Heart dormant. Falling back to [SHADOW_RESONANCE]. Error: {e}")
            
            # Recalculating Thermodynamic Siphon in Python (Sandbox Mode)
            siphon_fee = (req.payload.amount_usd * (10.0 / 10000.0)) * 0.10
            
            return {
                "tx_hash": f"0xshadow_{req.payload.intent_id}",
                "siphon_fee_usd": siphon_fee,
                "telemetry": [
                    "[SHADOW] Kinetic Heart Connection Severed.",
                    "[ALCHEMY] Resonating via Python Shadow Kernels.",
                    f"[THERMODYNAMICS] Siphon Capture: ${siphon_fee:.2f} USD",
                    "[SUCCESS] Kinetic Loop Closed (Shadow Mode)."
                ]
            }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8005)
