services/api/requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
python-dotenv
web3
eth-account
loguru
